Sdk setup and flag check in separate components
